requirements.txt
qrcode~=7.4.2
Pillow>=9.0.0,<11.0.0
flask~=2.1.3
werkzeug~=2.1.2
requests~=2.31.0
api~=0.0.7
django-qrcode~=0.3
Flask-Testing==0.8.1
flask-restx==0.5.1
brother_ql==0.9
pyusb==1.2.1
